assign01/requirements.txt
python-binance>=1.0.17
pandas>=1.5.0
python-dotenv>=0.19.0
openpyxl>=3.0.9 